tests: add more unit tests on logic expr parsing.
